+++ CHANGELOG.md
@@ -2176,3 +2176,4 @@ Alternately, the _Daily Diary of my Programming Dreams_.
 * **java**
 
   1. Added [fiddles/java/fiddle-0005-j2v8Hello](fiddles/java/fiddle-0005-j2v8Hello)
+* Added [fiddles/node/fiddle-0038-ZmqReqRep](fiddles/node/fiddle-0038-ZmqReqRep)